feat: add sample rule when started server
