+++ package.json
@@ -50,6 +50,7 @@
     "pino": "^8.18.0",
     "pino-http": "^9.0.0",
     "pino-pretty": "^10.3.1",
+    "tiny-invariant": "^1.3.1",
     "uuid": "^9.0.1",
     "zod": "^3.22.4",
     "zod-validation-error": "^3.0.0"
